Remove timestamp from the public API
kvzrtp has internal mechanisms for keeping track of time and this
timestamp was actually not used because its accuracy is questionable
AND it had no reference clock.

This change makes the library also easier to use because now user
only has to push data to the library and doesn't have to keep track of
time.