Fix Hello/HelloACK messages
Now they should work on both Linux and Windows and Hello messages
contain all the information required by the specification.

The parsing of algorithms is still missing but because kvzRTP's
implementation is only going to provide the mandatory algorithms,
the lists can be omitted and as long as we're only communicating with
ourselves, the parsing can be omitted too.

At some point in the future I'll add proper parsing and remove the GNU
extension from ZRTP to make it truly cross-platform and compatible with
other RTP libraries supporting ZRTP