Create new sender/receiver classes that are totally separate
The previous architectural design was starting show its limitations,
one of the most limiting factor being that readers and writers had
their own connection objects which caused some problems with hole
punching.

This is the first commit of a whole new architecture where the top
object is RTP session which contains media streams which in turn contain
an RTP sender and a receiver. This design choice allows the kvzRTP to use
one socket for both the receiver and the sender and thus mitigates the
hole punching problem that appears on Windows.