Don't store session objects to RTP context
Let user create as many sessions as wanted with the same IP.
kvzRTP really does not care how many sessions with the same IP is
going on so forcing only one session per IP made no sense.

Now, however, the burden of releasing the session object is on the
caller when the application exits.