Increase the RTP frame delay to 100ms
34ms is too strict a deadline for some use cases. By increasing the
default frame delay to 100ms does not hurt the cases where the frames
are anyway received within 34 milliseconds but it also allows users
to stream at high resolutions/low QP values without recompling the library.

Making this adjustment dynamic based on timestamps of the packets is
a planned future task but this shall act as a bandaid until then.