Fix race condition at the start up of system call dispatcher
There is a possibility that the dispatcher thread is started before
its state is initialized which causes it to appear as inactive when
it tries to enter the send loop and actually exists early.

This is a clear design flaw in the runner class but for the time
being this disgusting fix will have to suffice.