Improve Live555 latency test code
It seems that Live555 wants to be fed more than one frame at a time
and forcing it to take only one frame, send it and receive it back
doesn't seem possible.

The code is changed so that Live555 streams at low FPS and saves
each send time of a frame to a hashmap from which that timestamp
is later fetched when the frame is received.

The frame size is used as a key because it is the most reliable
way to identify incoming frames.