Rewrite FFmpeg's latency test code
Live555 latency test code requires us to change the design for
FFmpeg and uvgRTP too. Now the test code streams at 30 FPS and
calculates an average latency value for all frames, for intra frames
and for inter frames.

The presence of start codes and frame merging (probabyl due to presence
of SPS et al) makes the usage of frame size as key a little trickier.