Store pointer to the datagram in the rtp_frame structure
Auxiliary handlers may need access to the original block of memory
(such as SRTP) so instead of relaying that parameter to every handler,
store pointer to the block inside the RTP frame. The pointer is valid
as long as the packet is being processed and the last handler should
set it to nullptr.

Some changes to the public frame should be made because a lot of
the fields in the frames used by uvgRTP are not important outside
the library.