Fix holepuncher functionality for receiver-only streams
The previous assumption was that the stream utilizing holepuncher
would have already sent a datagram and would then switch just receiving
which is why the holepuncher did not send anything if the media stream
had not (yet) sent anything.

This does not, however, work for those streams that only receive packets
as they would never call push_frame() and thus never update the last_dgram_sent_
of the holepuncher, essentially making holepuncher disabled for the stream.