build: Build support for Linux g++ compilers not supporting c++17
This pull request allows the compilation of uvgRTP on CentOS7 where
c++17 and the getrandom() functions aren't available.

The CMakeLists.txt checks the c++ compiler standard conformance and
adjusts the gcc compilation directive -std accordingly.
It checks also the availability of the getrandom() function and creates
the HAVE_GETRANDOM preprocessor definition if this function exists.

If the getrandom() function doesn't exist, the equivalent syscall
replaces it.

If the c++ compiler is a c++17 or higher one, and the getrandom() is
available, the original code is left unchanged