build: merge polusto changes and test on CentOS7 and CentOS8
I tested the polusto changes on CentOS7 (c++11 g++ compiler, no
getrandom()) and CentOS8 stream (c++17 g++ compiler and getrandom()) :
 it doesn't compile anymore. The -std=c++11 or -std=c++17 aren't added
to the compile options in spite of the set(CMAKE_CXX_STANDARD 17)
directive.

I had to revert, for UNIX target, the computation of the compiler
definitions and options to the one of my original pull request.

Therefore, the uvgrtp library compiles successfully on CentOS7 and
CentOS8