formats: Use a non-static common packet_handler-function for h264/5/6
This commit removes most of the duplicate code left in the project
equaling to around 400 LOC. This was done by moving packet_handler
among few other functions to parent class h26x while making use of
virtualization to customize necessary components into child classes
h264, h265 and h266.

The virtualization however prevented the use of static functions,
which in turn required that the packet_dispatcher was modified to
support classes directly via std::function parameters.

This commit has only been tested with h265 and not with h264 or h266.

There may have still been a problem with appending of the start codes
for h264. This commit should fix that issue since now it is done in the
same way as other formats.