common: User ring buffer in packet dispatch
The ring buffer doesn't serve any function yet, but it will be used as
a buffer between receiver and processor thread. It seems to be working
if I tested it correctly. I also fixed a bug two commits ago where I
used buffer size instead of the actual buffer as a pointer.