common: Fix two off by one errors in receive flow
This fixes (pretty inconsequential bugs):

1) The first packet was written
to index 1, which meant that processing tried to process an empty
packet at the start of the program.

2) The last packet was never processed and the processing always
waited until the next packet had arrived. Not a major issue, but
good to have it fixed.

Also reduced the number of prints to make test output nicer.