common: Fix issues with fps flag
This fixes the synchronization issue with fragment pacing and removes
any potential buffers at the start of the stream. The fragment pacing
allocated time is reduced to 80% of frame time to allow for more
flexibility with fps enforcement.

The maximum fps wait was increased to frame interval, since smaller
wait seemed to cause issues.