formats: Fix h26x fragment reception with sequence roll-over
This commit fixes a bug where the frame is not reconstructed correctly
in case the sequence number rolls over inside the frame fragments.

The previous solution had the bug of always assuming that the std::set
was in correct order for the frame fragments, but this is not the case
if the seq rollover happens. The solution here is to use a separate
iterator for received fragments that starts at start fragment, not the
beginning of std::set and ends when start is found again.

# Conflicts:
#	src/formats/h26x.cc